+++ tutorial/src/CHANGELOG.md
@@ -51,3 +51,8 @@ exposed `hyperion_exposed` variable in visualization tool so that user can progr
 - support saving html with ctrl-S or cmd-S in GUI
 - update readme to include all the parameters
 - support sinter decoder inside mwpf
+
+## 0.2.6
+
+- add feature of constantly trying to solve primal problem even though there are still relaxers; this is good for decoding accuracy while still maintain the same level of time
+- release GIL during the decoding process